backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
python-decouple
whitenoise[bolt]
dj-database-url
gunicorn
psycopg2-binary
axios
